Split NED into topology and topobathy parts. This should help to alleviate the overlapping nature of the files doesn't cause too many problems.
